feat(one-login-auth-token): validates auth token claims and user sub (#83)
* feat(one-login-auth-token): validates auth token claims and user sub

---------

+++ src/main/java/gov/cabinetofice/gapuserservice/dto/IdTokenDto.java
@@ -14,8 +14,8 @@ public class IdTokenDto {
     private String at_hash;
     private String iss;
     private String aud;
-    private int exp;
-    private int iat;
+    private long exp;
+    private long iat;
     private String vot;
     private String vtm;
     private String sid;
